# Serialise asset graph as json